Update lib/merit/controller_extensions.rb
In line 48 of lib/merit/controller_extensions.rb, target_id is being 
evaluated by a regular expression. However, if target_id happens to 
be of type Fixnum or something other than a string type, nil will be 
returned. As nil is falsey, !(nil) evaluates to true, which can lead 
to the statement on line 49 being run undesirably. My change is a 
simple one: cast target_id as a string so it can be properly 
evaluated by the regular expression. This should eliminate the fringe 
case I encountered. Thanks.