Fix typo in active_record/merit_generator.rb
Line #16 was referring to the migration_template source as 'add_merit_fields_to_model.rb'.
This source does not exist as the actual extension is `.erb`. This caused the task
to fail as there is no template with the name 'add_merit_fields_to_model.rb'.

Resolving: https://github.com/merit-gem/merit/issues/344